main/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certifi==2022.12.7
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
commonmark==0.9.1
cryptography==43.0.1
Deprecated==1.2.14
Flask==1.1.2
Flask-CeleryExt==0.4.3
gitdb==4.0.11
GitPython==3.1.43
gunicorn==20.0.4
hiredis==3.0.0
idna==2.5
itsdangerous==0.24
Jinja2==2.11.3
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==1.1.1
mdurl==0.1.2
more-itertools==10.5.0
packaging==24.1
PatrowlEnginesUtils==1.2.0
prompt_toolkit==3.0.48
psutil==6.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGithub==2.1.0.post0
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ulid==1.1.0
PyYAML==6.0.2
redis==5.1.1
redis-om==0.3.2
requests==2.32.3
rich==13.8.1
setuptools==72.2.0
six==1.16.0
smmap==5.0.1
types-cffi==1.16.0.20240331
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==75.2.0.20241019
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.0.3
wrapt==1.16.0
